+++ crates/pop-common/src/git.rs
@@ -3,7 +3,8 @@
 use crate::{errors::Error, APP_USER_AGENT};
 use anyhow::Result;
 use git2::{
-	build::RepoBuilder, FetchOptions, IndexAddOption, RemoteCallbacks, Repository, ResetType,
+	build::RepoBuilder, FetchOptions, IndexAddOption, RemoteCallbacks, Repository as GitRepository,
+	ResetType,
 };
 use git2_credentials::CredentialHandler;
 use regex::Regex;
@@ -35,7 +36,7 @@ impl Git {
 		Ok(())
 	}
 
-	fn ssh_clone(url: &Url, working_dir: &Path) -> Result<Repository> {
+	fn ssh_clone(url: &Url, working_dir: &Path) -> Result<GitRepository> {
 		let ssh_url = GitHub::convert_to_ssh_url(url);
 		// Prepare callback and fetch options.
 		let mut fo = FetchOptions::new();
@@ -58,7 +59,7 @@ impl Git {
 		target: &Path,
 		tag_version: Option<String>,
 	) -> Result<Option<String>> {
-		let repo = match Repository::clone(url, target) {
+		let repo = match GitRepository::clone(url, target) {
 			Ok(repo) => repo,
 			Err(_e) => Self::ssh_clone_and_degit(
 				url::Url::parse(url).map_err(|err| Error::from(err))?,
@@ -92,7 +93,7 @@ impl Git {
 	}
 
 	/// For users that have ssh configuration for cloning repositories.
-	fn ssh_clone_and_degit(url: Url, target: &Path) -> Result<Repository> {
+	fn ssh_clone_and_degit(url: Url, target: &Path) -> Result<GitRepository> {
 		let ssh_url = GitHub::convert_to_ssh_url(&url);
 		// Prepare callback and fetch options.
 		let mut fo = FetchOptions::new();
@@ -118,7 +119,7 @@ impl Git {
 	}
 
 	/// Fetch the latest release from a repository
-	fn fetch_latest_tag(repo: &Repository) -> Option<String> {
+	fn fetch_latest_tag(repo: &GitRepository) -> Option<String> {
 		let version_reg = Regex::new(r"v\d+\.\d+\.\d+").expect("Valid regex");
 		let tags = repo.tag_names(None).ok()?;
 		// Start from latest tags
@@ -139,7 +140,7 @@ impl Git {
 	/// * `target` - location where the parachain will be created.
 	/// * `message` - message for first commit.
 	pub fn git_init(target: &Path, message: &str) -> Result<(), git2::Error> {
-		let repo = Repository::init(target)?;
+		let repo = GitRepository::init(target)?;
 		let signature = repo.signature()?;
 
 		let mut index = repo.index()?;
@@ -276,6 +277,41 @@ pub struct Release {
 	pub name: String,
 	pub prerelease: bool,
 	pub commit: Option<String>,
+}
+
+/// A descriptor of a remote repository.
+#[derive(Debug, PartialEq)]
+pub struct Repository {
+	/// The url of the repository.
+	pub url: Url,
+	/// If applicable, the branch or tag to be used.
+	pub reference: Option<String>,
+	/// The name of a package within the repository. Defaults to the repository name.
+	pub package: String,
+}
+
+impl Repository {
+	/// Parses a url in the form of https://github.com/org/repository?package#tag into its component parts.
+	///
+	/// # Arguments
+	/// * `url` - The url to be parsed.
+	pub fn parse(url: &str) -> Result<Self, Error> {
+		let url = Url::parse(url)?;
+		let package = url.query();
+		let reference = url.fragment().map(|f| f.to_string());
+
+		let mut url = url.clone();
+		url.set_query(None);
+		url.set_fragment(None);
+
+		let package = match package {
+			Some(b) => b,
+			None => crate::GitHub::name(&url)?,
+		}
+		.to_string();
+
+		Ok(Self { url, reference, package })
+	}
 }
 
 #[cfg(test)]
