+++ crates/pop-parachains/src/errors.rs
@@ -9,8 +9,6 @@ pub enum Error {
 	Aborted,
 	#[error("Anyhow error: {0}")]
 	AnyhowError(#[from] anyhow::Error),
-	#[error("Archive error: {0}")]
-	ArchiveError(String),
 	#[error("{0}")]
 	CommonError(#[from] pop_common::Error),
 	#[error("Configuration error: {0}")]
@@ -19,8 +17,6 @@ pub enum Error {
 	CurrentDirAccess,
 	#[error("Failed to parse the endowment value")]
 	EndowmentError,
-	#[error("HTTP error: {0}")]
-	HttpError(#[from] reqwest::Error),
 	#[error("IO error: {0}")]
 	IO(#[from] std::io::Error),
 	#[error("Missing binary: {0}")]
@@ -33,18 +29,14 @@ pub enum Error {
 	OrchestratorError(#[from] OrchestratorError),
 	#[error("Failed to create pallet directory")]
 	PalletDirCreation,
-	#[error("ParseError error: {0}")]
-	ParseError(#[from] url::ParseError),
 	#[error("Failed to execute rustfmt")]
 	RustfmtError(std::io::Error),
 	#[error("Template error: {0}")]
-	TemplateError(#[from] pop_common::templates::Error),
+	SourcingError(#[from] pop_common::sourcing::Error),
 	#[error("Toml error: {0}")]
 	TomlError(#[from] toml_edit::de::Error),
 	#[error("Unsupported command: {0}")]
 	UnsupportedCommand(String),
 	#[error("Failed to locate the workspace")]
 	WorkspaceLocate,
-	#[error("Unsupported platform: {arch} {os}")]
-	UnsupportedPlatform { arch: &'static str, os: &'static str },
 }
